requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
django-filter
flake8
isort
flake8-quotes